+++ src/components/icons/ListViewIcon.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+export const ListViewIcon: React.FC<React.SVGProps<SVGSVGElement>> = (
+  props,
+) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 -960 960 960"
+    fill="currentColor"
+    aria-hidden
+    focusable="false"
+    {...props}
+  >
+    <path d="M280-600v-80h480v80H280Zm0 160v-80h480v80H280Zm0 160v-80h480v80H280ZM200-600q-17 0-28.5-11.5T160-640q0-17 11.5-28.5T200-680q17 0 28.5 11.5T240-640q0 17-11.5 28.5T200-600Zm0 160q-17 0-28.5-11.5T160-480q0-17 11.5-28.5T200-520q17 0 28.5 11.5T240-480q0 17-11.5 28.5T200-440Zm0 160q-17 0-28.5-11.5T160-320q0-17 11.5-28.5T200-360q17 0 28.5 11.5T240-320q0 17-11.5 28.5T200-280Z" />
+  </svg>
+);