+++ src/components/icons/index.ts
@@ -1,10 +1,12 @@
 export { AnalyticsIcon } from "./AnalyticsIcon";
+export { BoardViewIcon } from "./BoardViewIcon";
 export { CardDragIcon } from "./CardDragIcon";
 export { CloseIcon } from "./CloseIcon";
 export { DensityLargeIcon } from "./DensityLargeIcon";
 export { DensityMediumIcon } from "./DensityMediumIcon";
 export { DensitySmallIcon } from "./DensitySmallIcon";
 export { DragIndicatorIcon } from "./DragIndicatorIcon";
+export { ListViewIcon } from "./ListViewIcon";
 export { OwlIcon } from "./OwlIcon";
 export { RedoIcon } from "./RedoIcon";
 export { SettingsGearIcon } from "./SettingsGearIcon";